Admin module fixes

+++ backend/src/site-admin/site-admin.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from "@nestjs/common";
 import { SiteAdminRepository } from './site-admin.repository';
 import { User } from '../user/entities/user.entity';
 import { SiteAdmin } from 'src/site-admin/entities/site-admin.entity';
@@ -8,6 +8,10 @@ export class SiteAdminService {
   constructor(private readonly siteAdminRepository: SiteAdminRepository) {}
 
   async addSiteAdmin(user: User): Promise<SiteAdmin> {
+    const isAdmin = await this.isUserSiteAdmin(user.id);
+    if (isAdmin) {
+      throw new BadRequestException('This user already has admin privilege');
+    }
     return await this.siteAdminRepository.create(user);
   }
 
@@ -17,6 +21,10 @@ export class SiteAdminService {
       throw new NotFoundException(
         `Site admin with user ID ${userId} not found`,
       );
+    }
+    const isAdmin = await this.isUserSiteAdmin(siteAdmin.id);
+    if (!isAdmin) {
+      throw new BadRequestException('This user has no admin privilege');
     }
     await this.siteAdminRepository.delete(siteAdmin.id);
   }
